Add multi-zone support for egressservice
This commit adds multi-zone support for egressservice with OVN-IC.
It introduces the usage of logical router static routes which are
required for redirecting traffic coming from PODs hosted in remote zones.

With IC disabled, for every matching pod IP, we create a logical router policy
with egressservice host mgmt IP as the next hop.

With IC enabled the following happens:
 - for matching pods hosted in the local zone:
   - if the egressservice host is in the local zone, create a logical router policy
     with the nexthop set to the egressservice hosts mgmt IP, just like when IC is
     disabled
   - if the egressservice host is in the remote zone, create a logical router policy
     with the nexthop set to the egressservice hosts transit switch IP
 - for matching pods hosted in a remote zone:
   - if the egressservice host is in the local zone, create a logical router
     static route to direct the packets from remote zone to the egressservice host
     mgmt IP
   - if the egressservice host is not in the local zone, do nothing

Signed-off-by: Patryk Diak <pdiak@redhat.com>