fix race condition in hybrid overlay.
at the beginning of AddNode() we checked if the hybrid annotation had
been set on the node being added as a quick way to determine if we
should continue. It turns out we also set the nodes drIP to whatever
the annotation of the added node is.

Turns out this does not effect the operation too much because we never
change the flows associated with the drIP except when nodeName being
added is equal to the nodeName of the controller but the internal
representation is wrong without this patch.

Signed-off-by: Jacob Tanenbaum <jtanenba@redhat.com>